implement base validation for orders
