+++ backintime/broker/broker.py
@@ -11,14 +11,16 @@ from .base import (
     LimitOrderInfo,
     StrategyOrders,
     StrategyOrderInfo,
-    validate_market_order,
-    validate_limit_order,
-    validate_take_profit_factory,
-    validate_stop_loss_factory,
     BrokerException,
     OrderSubmissionError,
     OrderCancellationError,
     AbstractBroker
+)
+from .validators import (
+    validate_market_order,
+    validate_limit_order,
+    validate_take_profit_factory,
+    validate_stop_loss_factory
 )
 from .orders import (
     OrderSide,
